app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["aarizona <ar.ariz.arizona@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.5"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
celery = {extras = ["asyncio", "redis"], version = "^5.4.0"}
redis = "^5.2.1"
gunicorn = "^23.0.0"
python-telegram-bot = "^21.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
